backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
pywin32==306
Werkzeug==3.0.1
